Re #18484 Use Instrument 2.0 in SaveCalFile

+++ Framework/Algorithms/src/GetDetectorOffsets.cpp
@@ -122,7 +122,7 @@ void GetDetectorOffsets::exec() {
   Progress prog(this, 0, 1.0, nspec);
   auto &spectrumInfo = maskWS->mutableSpectrumInfo();
   PARALLEL_FOR_IF(Kernel::threadSafe(*inputW))
-  for (size_t wi = 0; wi < nspec; ++wi) {
+  for (int64_t wi = 0; wi < static_cast<int64_t>(nspec); ++wi) {
     PARALLEL_START_INTERUPT_REGION
     // Fit the peak
     double offset = fitSpectra(wi, isAbsolute);
